Refactor render surface helpers
