+++ docs/code_flow.md
@@ -12,7 +12,7 @@ Describe how a `totem run` execution traverses configuration services, the runti
 
 ## Technical Approach
 
-Represent each execution stage as a node in a Mermaid flowchart. Colour code orchestration components, service layers, inputs, and outputs so reviewers can trace transitions. The diagram captures call sequencing between the CLI, configuration registry, runtime loop, app routing, peripheral managers, and display drivers. The goal is to surface every point where the runtime crosses a service boundary or hardware interface. The frame composition stage includes merge-strategy selection and cached surface reuse.
+Represent each execution stage as a node in a Mermaid flowchart. Colour code orchestration components, service layers, inputs, and outputs so reviewers can trace transitions. The diagram captures call sequencing between the CLI, configuration registry, runtime loop, app routing, peripheral managers, and display drivers. The goal is to surface every point where the runtime crosses a service boundary or hardware interface. The frame composition stage includes merge-strategy selection and cached surface reuse handled by the RendererSurfaceCache and SurfaceMerger helpers.
 
 ## Flow Diagram
 
@@ -36,7 +36,9 @@ flowchart LR
         Loop["GameLoop Service\n(heart.runtime.game_loop.GameLoop)"]
         AppRouter["AppController / Mode Router"]
         ModeServices["Mode Services & Renderers"]
+        SurfaceCache["Surface Cache\n(RendererSurfaceCache)"]
         FrameComposer["Frame Composer\n(surface merge + strategy selection)"]
+        SurfaceMerger["Surface Merger\n(SurfaceMerger)"]
     end
 
     subgraph Inputs["Peripheral & Signal Services"]
@@ -64,8 +66,9 @@ flowchart LR
     CLI --> Registry --> Configurer --> Loop
     Configurer --> AppRouter
     Loop --> AppRouter
-    Loop --> FrameComposer
-    AppRouter --> ModeServices --> FrameComposer --> DisplaySvc
+    Loop --> SurfaceCache --> FrameComposer --> SurfaceMerger
+    AppRouter --> ModeServices --> SurfaceCache
+    SurfaceMerger --> DisplaySvc
     DisplaySvc --> LocalScreen
     DisplaySvc --> Capture --> DeviceBridge --> LedMatrix
     Capture --> AverageMirror --> SingleLED
@@ -77,7 +80,7 @@ flowchart LR
     RxScheduler --> HeartRate --> AppRouter
     RxScheduler --> PhoneText --> AppRouter
 
-    class CLI,Registry,Configurer,ModeServices,FrameComposer service;
+    class CLI,Registry,Configurer,ModeServices,SurfaceCache,FrameComposer,SurfaceMerger service;
     class Loop,AppRouter orchestrator;
     class PeripheralMgr,RxScheduler,Switch,Gamepad,Sensors,HeartRate,PhoneText input;
     class DisplaySvc,LocalScreen,Capture,DeviceBridge,LedMatrix,AverageMirror,SingleLED output;
